requirements.txt
streamlit
pyyaml
python-dotenv
openai
sentence-transformers
tqdm
pandas
datasets
rank-bm25
nltk